Fortini-Engine-Windows-Package/requirements.txt
PyQt6>=6.5.0
PyOpenGL>=3.1.5
numpy>=1.24.0
pygame>=2.1.0
PyGLM>=2.5.0
